requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
uvicorn==0.30.6
crewai==0.121.0
pydantic==2.9.0
crewai-tools==0.45.0
pandas==2.2.3
streamlit==1.45.1